requirements.txt
absl-py==1.4.0
accelerate==0.23.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.2
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
cityscapesScripts==2.2.2
click==8.1.7
cmake==3.27.2
coloredlogs==15.0.1
contourpy==1.1.0
cycler==0.11.0
Cython==3.0.2
datasets==2.14.4
decorator==5.1.1
deepspeed==0.10.3
diffdist==0.1
dill==0.3.7
docker-pycreds==0.4.0
einops==0.7.0
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.104.0
ffmpy==0.3.1
filelock==3.12.2
fonttools==4.42.0
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gradio==3.42.0
gradio_client==0.5.0
grpcio==1.57.0
h11==0.14.0
h5py==3.9.0
hjson==3.1.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
imageio==2.31.1
imgaug==0.4.0
importlib-metadata==6.8.0
importlib-resources==6.0.1
infinibatch==0.1.1
iopath==0.1.10
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json-tricks==3.17.3
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
kornia==0.6.8
lazy_loader==0.3
lightning-utilities==0.9.0
lit==16.0.6
llvmlite==0.41.1
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mup==1.0.0
networkx==3.1
ninja==1.11.1.1
nltk==3.8.1
numba==0.58.1
numpy==1.23.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
openai-whisper==20230918
opencv-python==4.8.1.78
orjson==3.9.10
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
plotly==5.16.0
portalocker==2.8.2
prefetch-generator==1.0.3
prompt-toolkit==3.0.39
protobuf==4.24.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pydantic==1.10.13
pydub==0.25.1
Pygments==2.16.1
pyparsing==3.0.9
pyquaternion==0.9.9
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.6
rsa==4.9
safetensors==0.3.2
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.11.1
seaborn==0.12.2
SEEM @ git+https://github.com/UX-Decoder/Segment-Everything-Everywhere-All-At-Once.git@33f2c898fdc8d7c95dda014a4b9ebe4e413dbb2b
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
Semantic-SAM @ git+https://github.com/UX-Decoder/Semantic-SAM.git@0b1d10a87e7a15183ac528e70a646bf07a2df397
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.32.0
setproctitle==1.3.3
Shapely==1.8.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.8.12
tiktoken==0.3.3
timm==0.4.12
tokenizers==0.14.1
toolz==0.12.0
torch==2.0.1
torchmetrics==1.2.0
torchvision==0.15.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.34.0
triton==2.0.0
typing==3.7.4.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.2
vision-datasets==0.2.2
wandb==0.15.12
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.7
xxhash==3.3.0
yacs==0.1.8
yarl==1.9.2
zipp==3.16.2
